Add MethodHandleTest
This is a simple test case that illustrates the use of MethodHandles
with generic intermediaries that don't know about any specific
DynamicObject type, and therefore don't have direct access to the
methods on those types.

This is intended for use cases where a generic class (e.g. a generic
queue client) is to be developed that can work with arbitrary
DynamicObject types, while still attaching relevant metadata (which
requires calling builder methods). Since it is not possible to use
inheritance to extract the common metadata builder methods into a
separate interface, this technique is suggested as a type-safe
alternative.