Change return type of DynamicObject#getMap to a raw Map
The old wildcard return type of Map<?, ?> turns out to have poor
ergonomics; it's essentially useless and requires downcasting for many
methods to be callable (e.g. getOrDefault), and it certainly doesn't
provide any additional type safety. The wildcard return type can save a
cast if the returned map is homogeneous, but that is almost never the case
with a DynamicObject map.