Fix collection conversion bug
From clojure.org/transients:

"Note in particular that transients are not designed to be bashed
in-place.  You must capture and use the return value in the next call."

This bug was basically causing all maps larger than 2^n elements, where n
is a very small number, to break. (Lists do not appear to have been
affected, but they now comply with the above instructions anyway.)