EdnTranslator: Add Writer support
This adds a new method, EdnTranslator#write(T, Writer), that permits
unbuffered serialization directly to a writer. The default implementation
is given in terms of the prior write method that returns a String; that
method has also gained a default implementation that delegates to the
Writer-based method. The idea here is that it should never be necessary to
implement both methods.