Establish a built-in default reader
As a matter of principle, DynamicObject has always been capable of
dealing with unknown fields, not by silently dropping them but by
including them in the deserialized object in such a way they can be
serialized later. However, unknown reader tags have always been an
exception to this: the default behavior is to throw an exception as soon
as an unknown reader tag is encountered, and to make matters worse
DynamicObject has not (until very recently) exposed the Edn reader
option to specify a default reader.

After giving the matter some thought, I have decided that this is
entirely too reminiscent of the broken and inflexible defaults of tools
like Jackson. A far more sensible default behavior is to simply capture
both the tag and the element and preserve them in such a way that they
can be transparently re-serialized later. For the vast majority of use
cases, this is the solution that makes the most sense. A built-in
default reader saves users the trouble of having to create their own
default reader, and the provided Unknown class prevents the
proliferation of non-interoperable container types.