Support custom keys on builder methods
This adds @Key annotation support to builder methods, which can now have
arbitrary names (previously they had to have the same name as their
corresponding getter method). The intention of this change is to support
alternate naming conventions for builder methods, like setField,
withField, assocField, and so on.

Note that metadata builders currently require a @Meta annotation if they
are to have a custom name. (If a custom key is not specified, only the
getter requires such an annotation.)