Implement a faster validate() API
This improves validate() performance by avoiding reflection in the actual
validate() fast path. Instead, we perform all reflection on the first
invocation of validate(), cache the results (as a tree of lambdas), and bind
them to the validate() invokedynamic call site.

The performance improvements are quite large - when applied to a certain
real-world application with fairly expensive custom validate logic, the cost of
validate() calls still dropped by 2/3.

Behavior of validate() should be nearly unchanged, but it's possible that the
contents of exceptions when multiple validation issues are present might
change.