Fix FressianWriteHandler serialization bug
 ## Why is the change being made?
Clojure Fressian has a bug when serializing maps, see
here for details:
https://github.com/clojure/data.fressian/issues/14

This means users of DynamicObject have to serialize
dynamic objects at least twice
(serialize -> deserialize -> serialize)
to get a consistent serialization.

 ## What has changed to address the problem?
Changed FressianSerialization to create a PersistentHashMap if
we are writing a map of size 8. This ensures that every
serialization of a dynamic object returns the same value.

 ## How was this change tested?
Test-driven development. Reproduced the bug via a unit test.