Only remove tag if present
 ## Why is the change being made?
FressianSerialization#deregisterTag throws an exception if the tag is
not present in the binaryTagCache.
This is because the returned value from binaryTagCache.remove(type) will
be null, and fressianReadHandlers.remove(null) will throw an NPE.

 ## What has changed to address the problem?
Changed FressianSerialization's deregisterTag to check whether the tag
is present before removing it from fressianReadHandlers.
This keeps it consistent with EdnSerialization.

 ## How was this change tested?
Unit test added.